configured CMake and solved conversion warnings

+++ acquisition.hpp
@@ -7,14 +7,13 @@
 
 namespace img {
 
-std::vector<sf::Color>
-getPixels(const sf::Image& source) { // returns vector of pixels' color
+std::vector<sf::Color> getPixels(const sf::Image& source) {
   auto pixelArray{
       source.getPixelsPtr()}; // pointer to read-only color array (R,G,B,Alpha)
   auto size = source.getSize();
   auto n    = size.x * size.y; // number of pixels
   std::vector<sf::Color> pixels{};
-  for (int i{0}; i < n * 4; i += 4) {
+  for (unsigned int i{0}; i < n * 4; i += 4) {
     pixels.push_back({pixelArray[i], pixelArray[i + 1], pixelArray[i + 2],
                       pixelArray[i + 3]});
   }
@@ -23,37 +22,40 @@ getPixels(const sf::Image& source) { // returns vector of pixels' color
 }
 
 sf::Image resize(const sf::Image& source) { // bilinear interpolation
-  int N{64};                                // size of new image
+  unsigned int N{64};                       // size of new image
   auto sourceSize{source.getSize()};
   auto sourceX{sourceSize.x};
   auto sourceY{sourceSize.y};
 
-  float xConv{static_cast<float>(sourceX - 1) / (N - 1)};
-  float yConv{static_cast<float>(sourceY - 1) / (N - 1)}; // conversion factors
+  float xConv{static_cast<float>(sourceX - 1) / static_cast<float>(N - 1)};
+  float yConv{static_cast<float>(sourceY - 1)
+              / static_cast<float>(N - 1)}; // conversion factors
 
   sf::Image resized;
   resized.create(N, N); // sets resized size to NxN
 
-  for (int x{0}; x < N; ++x) {
-    for (int y{0}; y < N; ++y) { // coordinates in resized (to fill image)
-      float xF{x * xConv};
-      float yF{y * yConv}; // equivalent coordinates in source
+  for (unsigned int x{0}; x < N; ++x) {
+    for (unsigned int y{0}; y < N; ++y) {
+      // equivalent coordinates in source
+      float xF{static_cast<float>(x) * xConv};
+      float yF{static_cast<float>(y) * yConv};
+      // equivalent coordinates in source
 
-      int x1{static_cast<int>(xF)};
-      int y1{static_cast<int>(yF)}; // top left pixel
+      unsigned int x1{static_cast<unsigned int>(xF)};
+      unsigned int y1{static_cast<unsigned int>(yF)}; // top left pixel
 
-      int x2{std::min(x1 + 1, static_cast<int>(sourceX - 1))};
-      int y2{std::min(
-          y1 + 1, static_cast<int>(
-                      sourceY - 1))}; // down right pixel (accounting borders)
+      unsigned int x2{std::min(x1 + 1, static_cast<unsigned int>(sourceX - 1))};
+      unsigned int y2{std::min(y1 + 1, static_cast<unsigned int>(sourceY - 1))};
+      // down-right pixel
 
       sf::Color Q11{source.getPixel(x1, y1)};
       sf::Color Q21{source.getPixel(x2, y1)};
       sf::Color Q12{source.getPixel(x1, y2)};
       sf::Color Q22{source.getPixel(x2, y2)}; // gets colors from source
 
-      float xWeight{xF - x1};
-      float yWeight{yF - y1}; // linear interpolation weights
+      float xWeight{xF - static_cast<float>(x1)};
+      float yWeight{yF
+                    - static_cast<float>(y1)}; // linear interpolation weights
 
       auto linErp = [](sf::Uint8 c1, sf::Uint8 c2, float weight) {
         return static_cast<sf::Uint8>(c1 * (1 - weight) + c2 * weight);
@@ -81,7 +83,7 @@ std::vector<int> binaryConvert(sf::Image source) { // converts image to binary
   auto pixels{getPixels(source)};
   std::vector<int> binary{};
 
-  for (int i{0}; i < pixels.size(); ++i) {
+  for (long unsigned int i{0}; i < pixels.size(); ++i) {
     int colorAvg{
         static_cast<int>((pixels[i].r + pixels[i].b + pixels[i].g) / 3)};
     if (colorAvg >= 127) {
@@ -95,13 +97,13 @@ std::vector<int> binaryConvert(sf::Image source) { // converts image to binary
 
 sf::Image formImage(std::vector<int> binary) { // forms NxN image
   sf::Image result;
-  int N{64};
+  unsigned int N{64};
   result.create(N, N);
   // controlla che binary.size() abbia senso
 
-  int i{0};
-  for (int y{0}; y < N; ++y) {
-    for (int x{0}; x < N; ++x) {
+  long unsigned int i{0};
+  for (unsigned int y{0}; y < N; ++y) {
+    for (unsigned int x{0}; x < N; ++x) {
       if (binary[i] == -1) {
         result.setPixel(x, y, sf::Color::Black);
       } else {
